+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 1. [#130](https://github.com/InfluxCommunity/influxdb3-python/pull/130): Remove org parameters from the example code because It is not mandatory in Influxdb3
+2. [#139](https://github.com/InfluxCommunity/influxdb3-python/pull/139): Supports environment variables with the same name like other clients   
 
 ## 0.12.0 [2025-03-26]